==> 2-random_forest_analyses/2-RF_fit_models_per_tumor.R <==
library(caTools)
library(AUC)
library(caret)
library(randomForest)
library(doMC)
library(e1071)
library(tidyverse)
set.seed(1)

# This script is to obtain a feature ranking for each of the 4 subsets of
# features. the features are stored as separate files, one for each subset. I
# will use a partition of the train data as an input of the feature ranking.

# Import data -------------------------------------------------------------

# Set working directory
setwd("/DATA/users/k.bresser/ligandome_project/analysis_new/Ligandome_models")

# Read in train set
train.peptides <- read_tsv("./Output/test_train_tables_new/rf_train_peptides.tsv")

# Drop columns not needed for training and switch ligand column to factor
train.peptides %>% 
  mutate(ligand = factor(case_when(ligand == TRUE ~ "yes", TRUE ~ "no"), 
                         levels = c("yes", "no") )) %>% 
  dplyr::select(tumor, ligand, swissprot_id) -> train.peptides

# Sub-sample train set
train.peptides %>% 
  group_by(tumor, ligand) %>% 
  nest() %>% 
  ungroup() %>% 
  mutate(sample.size = c(4000,4000,4000, 2000, 2000, 2000) ) %>% 
#  mutate(sample.size = c(1000,1000,1000, 250, 250, 250) ) %>% 
  mutate(data = map2(data, sample.size, sample_n)) %>% 
  select(-sample.size) %>% 
  unnest(data) -> train.peptides


# Import features tables
paths <- paste0("./Output/feature_files/", list.files("./Output/feature_files/"))

list.files("./Output/feature_files/") %>% 
  str_split("_") %>% 
  map_chr(1) -> table.names

paths %>% 
  map(read_tsv) %>% 
  set_names(table.names) -> feature.tables

tibble(class = names(feature.tables),
       count = map_dbl(feature.tables, ncol) -1) %>% 
  ggplot(aes(x = fct_relevel(class, "UTR5", "CDS","UTR3","hsa"), y = log10(count), label = count))+
  geom_bar(stat = "identity")+
  geom_text(nudge_y = 0.1)+
  theme_classic()+
  theme(panel.grid.major.y = element_line())

ggsave("./Figs/number_features.pdf", units = "mm", width = 30, height = 36, scale = 2)


feature.tables %>% 
  enframe("classes", "values") %>% 
  mutate(values = map(values, inner_join, train.peptides, c("Entry" = "swissprot_id")),
         values = map(values, group_by, tumor),
         values = map(values, nest)) %>% 
  unnest(values) -> features.per.tumor




# random forest models ----------------------------------------------------


# Set up two class summary function
twoClassSummaryCustom <- function(data, lev = NULL, model = NULL) 
{
  lvls <- levels(data$obs)
  if(length(lvls) > 2)
    stop(paste("Your outcome has", length(lvls), "levels. The twoClassSummary() function isn't appropriate."))
  if(!all(levels(data[, "pred"]) == lvls)) 
    stop("levels of observed and predicted data do not match")
  rocAUC <- ModelMetrics::auc(ifelse(data$obs == lev[2], 0, 
                                     1), data[, lvls[1]])
  #  pred <- factor(ifelse(data[, "no"] > 0.5, "no", "yes"))
  #  pred <- factor(ifelse(data[, "yes"] > 0.99, "yes", "no"))
  
  out <- c(rocAUC,
           sensitivity(data[, "pred"], data[, "obs"], lev[1]),
           specificity(data[, "pred"], data[, "obs"], lev[2]),
           posPredValue(data[, "pred"], data[, "obs"], lev[1]))
  names(out) <- c("ROC", "Sens","Spec", "Prec")
  out
}


# Register cores for parallel processing
registerDoMC(10)

# Set the train control
trctrl <-  trainControl(method = "repeatedcv", number = 10, repeats = 1, 
                        summaryFunction = twoClassSummaryCustom, classProbs = TRUE, 
                        returnData = FALSE, sampling = "down", allowParallel = T)

# Define function to build formula from column names
get_formulas <- function(df){
  df %>% 
    select(-c(Entry, ligand)) %>% 
    names() %>% 
    paste( collapse = " + ") -> feats
  
  form <- paste("ligand", feats, sep = " ~ ")
  form <- as.formula(form)
  form
}


# Get the formula's and add to tibble
features.per.tumor %>% 
  mutate(formulas = map(data, get_formulas)) -> features.per.tumor
  


features.per.tumor %>% 
  mutate(rf.models = map2(formulas, 
                          data, 
                          ~train(.x , data = .y, 
                                 method = "rf",
                                 maximize = TRUE,
                                 trControl=trctrl,
                                 tuneGrid= data.frame(mtry = c(round(sqrt(ncol(.y)-1))))*1.5,
                                 importance = T, 
                                 nodesize = 2, 
                                 ntree = 5000,
                                 metric = "ROC"))) %>% 
  select(tumor, classes, rf.models) -> rf.models



write_rds(rf.models, "./Output/RF_per_tumor_new.rds")
